dev21w01a
Added a check for outlined products. Outlined products have a >0 opacity when unselected.

Attached `EventHandler` instance to constructor of `Configurator` opposed to spawning a new instance every time a click event occurs. This allows the two classes to interact with the same `config` object.

instance of `EventHandler.addProduct()` now removes incompatible products.

Clicks on the modal wrapper now passes the event to `close()` to determine if the modal is without a header.

Changed naming convention "startActive" to "enabled".